requirements.txt
django >= 2.1.5
djangorestframework >=3.9.1
coreapi >= 2.3.3
flake8 >= 3.6.0
